requirements.txt
absl-py==1.3.0
anyio==3.4.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bidict==0.22.0
bleach==4.1.0
cachetools==5.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.5.0
decorator==4.4.2
defusedxml==0.7.1
Django==4.1.6
dnspython==2.2.1
entrypoints==0.3
eventlet==0.33.2
expiringdict==1.2.2
fastapi==0.63.0
fasteners==0.16.3
filelock==3.9.0
Flask==2.0.2
Flask-SocketIO==5.1.1
flatbuffers==22.12.6
fonttools==4.28.5
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h11==0.12.0
h5py==3.7.0
huggingface-hub==0.12.0
idna==2.10
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==5.1.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.1
joblib==1.2.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.13.1
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.11.0
kiwisolver==1.3.2
Levenshtein==0.20.9
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
moviepy==1.0.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
nltk==3.8.1
notebook==6.4.7
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pd==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
proglog==0.1.10
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.6
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycparser==2.21
pydantic==1.8.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.9.0
pymongo==4.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-engineio==4.3.4
python-Levenshtein==0.20.9
python-socketio==5.7.2
python-version==0.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.2.2
QtPy==2.0.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.7.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sos==0.22.5
sqlparse==0.4.3
starlette==0.13.6
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.1.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.61.1
traitlets==5.1.0
transformers==4.26.0
typer==0.7.0
typing_extensions==4.4.0
tzlocal==2.1
urllib3==1.26.3
uvicorn==0.13.4
wasabi==1.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.14.1
xxhash==2.0.2
zipp==3.11.0
